Made new handles default for all shapes
