+++ packages/annotorious-openseadragon/test/index.html
@@ -41,10 +41,10 @@
       };
 
       // Ashmolean Museum Evans Architectural Plans HF/2
-      const IMAGE = 'https://iiif.bodleian.ox.ac.uk/iiif/image/af315e66-6a85-445b-9e26-012f729fc49c/info.json';
+      // const IMAGE = 'https://iiif.bodleian.ox.ac.uk/iiif/image/af315e66-6a85-445b-9e26-012f729fc49c/info.json';
 
       // Balliol College MS 238A
-      // const IMAGE = 'https://iiif.bodleian.ox.ac.uk/iiif/image/79bf8325-22fa-4696-afe5-7d827d84f393/info.json';
+      const IMAGE = 'https://iiif.bodleian.ox.ac.uk/iiif/image/79bf8325-22fa-4696-afe5-7d827d84f393/info.json';
 
       window.onload = function() {
         var viewer = OpenSeadragon({
