+++ packages/annotorious/src/annotation/editors/Handle.svelte
@@ -5,63 +5,90 @@
   export let x: number;
   export let y: number;
   export let scale: number;
-  export let radius: number = 30;
+  export let selected: Boolean | undefined = undefined;
 
   let touched = false;
 
-  const onPointerDown = (event: PointerEvent) => {
-    if (event.pointerType === 'touch')
+  const onPointerDown = (evt: PointerEvent) => {
+    if (evt.pointerType === 'touch')
       touched = true;
   }
 
   const onPointerUp = () =>
     touched = false;
 
-  $: handleSize = 10 / scale;
+  $: handleRadius = 4 / scale;
 </script>
 
 {#if isTouch}
-  <g class="a9s-touch-handle">
+  <circle 
+    cx={x}
+    cy={y}
+    r={2 * handleRadius} />
+{:else}
+  <g class={`a9s-handle ${$$props.class || ''}`.trim()}>
     <circle 
-      cx={x} 
-      cy={y} 
-      r={radius / scale}
-      class="a9s-touch-halo"
-      class:touched={touched}
+      class="a9s-handle-buffer"
+      cx={x}
+      cy={y}
+      r={handleRadius + (6 / scale)} 
+      on:pointerenter
+      on:pointerleave
       on:pointerdown
+      on:pointerdown={onPointerDown}
       on:pointerup
-      on:pointerdown={onPointerDown} 
       on:pointerup={onPointerUp} />
 
-    <rect 
-      class={`a9s-handle ${$$props.class || ''}`.trim()}
-      x={x - handleSize / 2} 
-      y={y - handleSize / 2} 
-      width={handleSize} 
-      height={handleSize}
-      on:pointerdown
-      on:pointerdown={onPointerDown} 
-      on:pointerup
-      on:pointerup={onPointerUp}  />
+    {#if selected}
+      <circle 
+        class="a9s-handle-selected"
+        cx={x} 
+        cy={y} 
+        r={handleRadius + (6 / scale)} />
+    {/if}
+
+    <circle 
+      class={`a9s-handle-dot${selected ? ' selected': ''}`}
+      cx={x}
+      cy={y}
+      r={handleRadius} />
   </g>
-{:else}
-  <rect 
-    class={`a9s-handle ${$$props.class || ''}`.trim()}
-    x={x - handleSize / 2} 
-    y={y - handleSize / 2} 
-    width={handleSize} 
-    height={handleSize} 
-    on:pointerdown
-    on:pointerup />
 {/if}
 
 <style>
-  .a9s-touch-halo {
+  .a9s-handle-buffer {
     fill: transparent;
-    stroke-width: 0;
   }
 
-  .a9s-touch-halo.touched {
+  .a9s-handle-dot {
+    fill: #fff;
+    pointer-events: none;
+    stroke: rgba(0, 0, 0, 0.35);
+    stroke-width: 1px;
+    vector-effect: non-scaling-stroke;
+  }
+
+  .a9s-handle-dot.selected {
+    fill: #1a1a1a;
+    stroke: none;
+  }
+
+  .a9s-handle-selected {
+    animation: dash-rotate 350ms linear infinite reverse;
     fill: rgba(255, 255, 255, 0.25);
+    stroke: rgba(0, 0, 0, 0.9);
+    stroke-dasharray: 2 2;
+    stroke-width: 1px;
+    pointer-events: none;
+    vector-effect: non-scaling-stroke;
+  }
+
+  @keyframes dash-rotate {
+    0% {
+      stroke-dashoffset: 0;
+    }
+    100% {
+      stroke-dashoffset: 4; /* Sum of dash + gap */
+    }
   }
 </style>
