+++ packages/annotorious/src/annotation/editors/editorsRegistry.ts
@@ -1,13 +1,12 @@
 import type { SvelteComponent } from 'svelte';
 import { ShapeType, type Shape } from '../../model';
 import { MultiPolygonEditor } from './multipolygon';
-import { BetterPolygonEditor, PolygonEditor } from './polygon';
+import { PolygonEditor } from './polygon';
 import { RectangleEditor } from './rectangle';
 
 const REGISTERED = new Map<ShapeType, typeof SvelteComponent>([
   [ShapeType.RECTANGLE, RectangleEditor as typeof SvelteComponent],
-  // [ShapeType.POLYGON, PolygonEditor as typeof SvelteComponent],
-  [ShapeType.POLYGON, BetterPolygonEditor as typeof SvelteComponent],
+  [ShapeType.POLYGON, PolygonEditor as typeof SvelteComponent],
   [ShapeType.MULTIPOLYGLON, MultiPolygonEditor as typeof SvelteComponent]
 ]);
