+++ packages/annotorious/src/annotation/editors/polygon/index.ts
@@ -1,2 +1,1 @@
 export { default as PolygonEditor } from './PolygonEditor.svelte';
-export { default as BetterPolygonEditor } from './BetterPolygonEditor.svelte';
